+++ src/plugins/contentRenderer/matrix/matrixContentRenderer.component.ts
@@ -1,9 +1,10 @@
 import {Component, ChangeDetectionStrategy, inject, ElementRef, Inject, Optional, ValueProvider, ViewChild, ViewContainerRef, OnDestroy, FactoryProvider, Injector, OnInit, ComponentRef} from '@angular/core';
+import {toObservable} from '@angular/core/rxjs-interop';
 import {RecursivePartial, extend} from '@jscrpt/common';
 import {Subscription} from 'rxjs';
 
 import {MatrixContentRenderer, MatrixContentRendererDefautTemplates, MatrixContentRendererOptions} from './matrixContentRenderer.interface';
-import {ContentRendererInnerStructure, DataLoader, DataResponse, Grid, GridContext, GridDataRowContext, GridPlugin, GridRowContext, MatrixGridMetadata, MetadataSelector, Paging, RowSelector} from '../../../interfaces';
+import {ContentRendererInnerStructure, DataLoader, DataResponse, Grid, GridContext, GridDataRowContext, GridPlugin, GridRowContext, MatrixGridColumn, MatrixGridMetadata, MetadataSelector, Paging, RowSelector} from '../../../interfaces';
 import {GridPluginInstances} from '../../../misc/types';
 import {CONTENT_RENDERER_INNER_STRUCTURE, CONTENT_RENDERER_OPTIONS, DEFAULT_OPTIONS, GRID_INSTANCE, GRID_PLUGIN_INSTANCES} from '../../../misc/tokens';
 import {CssGridDefaultTemplatesSAComponent} from './misc/components';
@@ -119,7 +120,7 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
     /**
      * Metadata selector currently used
      */
-    protected metadataSelector: MetadataSelector<MatrixGridMetadata>|undefined|null;
+    protected metadataSelector: MetadataSelector<MatrixGridMetadata<MatrixGridColumn>>|undefined|null;
 
     /**
      * Data loader currently used
@@ -140,6 +141,11 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
      * Subscription for metadata changes
      */
     protected metadataChangeSubscription: Subscription|undefined|null;
+
+    /**
+     * Subscription for data changes
+     */
+    protected dataChangeSubscription: Subscription|undefined|null;
 
     /**
      * Instance with default templates
@@ -205,7 +211,7 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
     }
 
     //######################### public methods - implementation of OnInit #########################
-    
+
     /**
      * Initialize component
      */
@@ -229,6 +235,9 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
     {
         this.metadataChangeSubscription?.unsubscribe();
         this.metadataChangeSubscription = null;
+
+        this.dataChangeSubscription?.unsubscribe();
+        this.dataChangeSubscription = null;
     }
 
     //######################### public methods - implementation of MatrixContentRenderer #########################
@@ -238,7 +247,7 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
      */
     public initialize(force: boolean): void
     {
-        const metadataSelector: MetadataSelector<MatrixGridMetadata> = this.gridPluginsSafe.MetadataSelector as MetadataSelector<MatrixGridMetadata>;
+        const metadataSelector: MetadataSelector<MatrixGridMetadata<MatrixGridColumn>> = this.gridPluginsSafe.MetadataSelector as MetadataSelector<MatrixGridMetadata<MatrixGridColumn>>;
 
         //metadata selector obtained and its different instance
         if(force || (this.metadataSelector && this.metadataSelector != metadataSelector))
@@ -262,6 +271,9 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
         //data loader obtained and its different instance
         if(force || (this.dataLoader && this.dataLoader != dataLoader))
         {
+            this.dataChangeSubscription?.unsubscribe();
+            this.dataChangeSubscription = null;
+
             this.dataLoader = null;
         }
 
@@ -269,6 +281,8 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
         if(!this.dataLoader)
         {
             this.dataLoader = dataLoader;
+
+            this.dataChangeSubscription = toObservable(this.dataLoader.result, {injector: this.injector}).subscribe(() => this.renderGridContainer());
         }
 
         const paging: Paging = this.gridPluginsSafe.Paging as Paging;
@@ -325,9 +339,15 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
     {
         //remove existing
         this.container.clear();
+        this.innerStructure.gridContainer = {renderableContent: null, view: null};
+        this.innerStructure.headerContainer = {renderableContent: null, view: null};
+        this.innerStructure.contentContainer = {renderableContent: null, view: null};
+        this.innerStructure.footerContainer = {renderableContent: null, view: null};
+        this.innerStructure.headerRowContainer = [];
+        this.innerStructure.contentRowContainer = [];
+        this.innerStructure.footerRowContainer = [];
 
-
-        this.innerStructure.gridContainer.view = this.container.createEmbeddedView(this.metadataSelector?.metadata?.gridContainer?.template ?? this.defaultsSafe.gridContainer, 
+        this.innerStructure.gridContainer.view = this.container.createEmbeddedView(this.metadataSelector?.metadata?.gridContainer?.template ?? this.defaultsSafe.gridContainer,
                                                                                    this.getGridContext(),
                                                                                    {
                                                                                        injector: this.createInjector(this.container.injector),
@@ -364,7 +384,9 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
                                                                                       {
                                                                                           injector: this.createInjector(viewContainer.injector),
                                                                                       });
-    
+
+        this.innerStructure.headerContainer.renderableContent = null;
+        this.innerStructure.headerRowContainer = [];
         this.innerStructure.headerContainer.view?.detectChanges();
 
         //header container is renderable
@@ -387,8 +409,17 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
                                                                                       {
                                                                                           injector: this.createInjector(viewContainer.injector),
                                                                                       });
-    
+
+        this.innerStructure.contentContainer.renderableContent = null;
+        this.innerStructure.contentRowContainer = [];
         this.innerStructure.contentContainer.view?.detectChanges();
+
+        //content container is renderable
+        if(this.innerStructure.contentContainer.renderableContent)
+        {
+            //render content rows
+            this.renderContentRowsContainers();
+        }
     }
 
     /**
@@ -403,8 +434,17 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
                                                                                       {
                                                                                           injector: this.createInjector(viewContainer.injector),
                                                                                       });
-    
+
+        this.innerStructure.footerContainer.renderableContent = null;
+        this.innerStructure.footerRowContainer = [];
         this.innerStructure.footerContainer.view?.detectChanges();
+
+        //footer container is renderable
+        if(this.innerStructure.footerContainer.renderableContent)
+        {
+            //render footer rows
+            this.renderFooterRowsContainers();
+        }
     }
 
     /**
@@ -418,27 +458,29 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
         viewContainer?.clear();
 
         this.innerStructure.headerRowContainer = [];
-        let index = 0;
+        const headerRowsTemplates = this.metadataSelector?.metadata?.headerRowContainer?.length ? this.metadataSelector?.metadata?.headerRowContainer : [{template: this.defaultsSafe.headerRowContainer, predicate: null}];
 
-        for(const headerRow of this.metadataSelector?.metadata?.headerRowContainer ?? [{template: this.defaultsSafe.headerRowContainer}])
+        for(let index = 0; index < headerRowsTemplates.length; index++)
         {
+            const context = this.getGridRowContext(index);
+            const headerRow = headerRowsTemplates[index];
+
+            //skip rendering of this row
+            if(headerRow.predicate && !headerRow.predicate(context))
+            {
+                continue;
+            }
+
+            this.innerStructure.headerRowContainer.push({renderableContent: null, view: null});
+
             this.innerStructure.headerRowContainer[index].view = viewContainer?.createEmbeddedView(headerRow.template,
-                                                                                                     this.getGridRowContext(index),
-                                                                                                     {
-                                                                                                         injector: this.createInjector(viewContainer.injector),
-                                                                                                     });
-        
+                                                                                                   context,
+                                                                                                   {
+                                                                                                       injector: this.createInjector(viewContainer.injector),
+                                                                                                   });
 
-            index++;
-            this.innerStructure.headerContainer.view?.detectChanges();
+            this.innerStructure.headerRowContainer[index].view?.detectChanges();
         }
-    }
-
-    /**
-     * Renders header row container
-     */
-    protected renderHeaderRowContainer(): void
-    {
     }
 
     /**
@@ -446,13 +488,65 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
      */
     protected renderContentRowsContainers(): void
     {
-    }
+        const viewContainer = this.innerStructure.contentContainer.renderableContent?.viewContainer;
 
-    /**
-     * Renders content row container
-     */
-    protected renderContentRowContainer(): void
-    {
+        //removing renderable content
+        viewContainer?.clear();
+
+        this.innerStructure.contentRowContainer = [];
+        const contentRowsTemplates = this.metadataSelector?.metadata?.contentRowContainer?.length ? this.metadataSelector?.metadata?.contentRowContainer : [{template: this.defaultsSafe.contentRowContainer, predicate: null}];
+
+        for(let datumIndex = 0; datumIndex < (this.dataLoader?.result().data?.length ?? 0); datumIndex++)
+        {
+            const datum = this.dataLoader?.result().data[datumIndex];
+            const context = this.getGridDataRowContext(datumIndex, datum);
+
+            this.innerStructure.contentRowContainer.push([]);
+
+            for(let index = 0; index < contentRowsTemplates.length; index++)
+            {
+                const contentRow = contentRowsTemplates[index];
+
+                //skip rendering of this row
+                if(contentRow.predicate && !contentRow.predicate(context))
+                {
+                    continue;
+                }
+
+                this.innerStructure.contentRowContainer[datumIndex].push({renderableContent: null, view: null});
+
+                this.innerStructure.contentRowContainer[datumIndex][index].view = viewContainer?.createEmbeddedView(contentRow.template,
+                                                                                                                    context,
+                                                                                                                    {
+                                                                                                                        injector: this.createInjector(viewContainer.injector),
+                                                                                                                    });
+
+                this.innerStructure.contentRowContainer[datumIndex][index].view?.detectChanges();
+
+                //TODO move into function
+                const columnViewContainer = this.innerStructure.contentRowContainer[datumIndex][index].renderableContent?.viewContainer;
+                const columns = this.metadataSelector?.metadata?.columns ?? [];
+
+                columnViewContainer?.clear();
+
+                for(const column of columns)
+                {
+                    if(!column.bodyTemplate)
+                    {
+                        throw new Error('MatrixContentRendererSAComponent: missing content (body) template!');
+                    }
+
+                    columnViewContainer?.createEmbeddedView(column.bodyTemplate,
+                                                            {
+                                                                ...context,
+                                                                metadata: column,
+                                                            },
+                                                            {
+                                                                injector: this.createInjector(columnViewContainer.injector),
+                                                            });
+                }
+            }
+        }
     }
 
     /**
@@ -460,13 +554,35 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
      */
     protected renderFooterRowsContainers(): void
     {
-    }
+        const viewContainer = this.innerStructure.footerContainer.renderableContent?.viewContainer;
 
-    /**
-     * Renders footer row container
-     */
-    protected renderFooterRowContainer(): void
-    {
+        //removing renderable content
+        viewContainer?.clear();
+
+        this.innerStructure.footerRowContainer = [];
+        const footerRowsTemplates = this.metadataSelector?.metadata?.footerRowContainer?.length ? this.metadataSelector?.metadata?.footerRowContainer : [{template: this.defaultsSafe.footerRowContainer, predicate: null}];
+
+        for(let index = 0; index < footerRowsTemplates.length; index++)
+        {
+            const context = this.getGridRowContext(index);
+            const footerRow = footerRowsTemplates[index];
+
+            //skip rendering of this row
+            if(footerRow.predicate && !footerRow.predicate(context))
+            {
+                continue;
+            }
+
+            this.innerStructure.footerRowContainer.push({renderableContent: null, view: null});
+
+            this.innerStructure.footerRowContainer[index].view = viewContainer?.createEmbeddedView(footerRow.template,
+                                                                                                   context,
+                                                                                                   {
+                                                                                                       injector: this.createInjector(viewContainer.injector),
+                                                                                                   });
+
+            this.innerStructure.footerRowContainer[index].view?.detectChanges();
+        }
     }
 
     /**
@@ -495,7 +611,6 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
         };
     }
 
-    //TODO: finish
     /**
      * Gets grid data row context
      * @param index - Current index to be used for creation of grid row context
@@ -517,8 +632,7 @@ export class MatrixContentRendererSAComponent implements MatrixContentRenderer, 
             get isSelected(): boolean
             {
                 return rowSelector?.isSelected(datum) ?? false;
-            },    
-
+            },
         };
     }