+++ src/plugins/rowSelector/index.ts
@@ -2,3 +2,5 @@ export * from './basic/basicRowSelector.component';
 export * from './basic/basicRowSelector.interface';
 export * from './limited/limitedRowSelector.component';
 export * from './limited/limitedRowSelector.interface';
+export * from './no/noRowSelector.component';
+export * from './no/noRowSelector.interface';